Card Scriptable Objects
I created Scriptable Objects for the four types of cards (Action, Deck Leader, Skill, and Monster).

By right clicking the project assets, you are able to create a new card scriptable object and change the values.

I created a Parent Script for Card Effects. In theory, we should be able to create more general effects and manipulate them to be more unique for each card.

I created a prefab for a Monster Card. By dragging a Monster Card Scriptable Object into the Card value in the Monster Card Display script, it will display the card's attributes on the visible object.