Card Stack Viewing
Created Scripts for: Deck, Extra Deck, Hand, Shield, Discard Pile.
They each have methods that allow them to send cards to each other.

Created the Card Stack View Manager.
This controls the player's ability to look into their Deck, Extra Deck, Shield, Discard Pile. When you click on a card in the view window, you will see a slightly more clear view of the card in the Card Information to the left. Will need to add more restriction to it to change cards face down in the screen to hide information in the Shield.